Address more comments

+++ src/core/sent_packet_metadata.c
@@ -62,8 +62,8 @@ QuicSentPacketPoolInitialize(
     _Inout_ QUIC_SENT_PACKET_POOL* Pool
     )
 {
-    for (size_t i = 0; i < ARRAYSIZE(Pool->Pools); i++) {
-        size_t PacketMetadataSize =
+    for (uint32_t i = 0; i < ARRAYSIZE(Pool->Pools); i++) {
+        uint32_t PacketMetadataSize =
             (i + 1) * sizeof(QUIC_SENT_FRAME_METADATA) +
             sizeof(QUIC_SENT_PACKET_METADATA);
